Added superscript number for cart
