+++ src/cart/index.html
@@ -10,7 +10,6 @@
 
     <link rel="stylesheet" href="/css/style.css" />
 
-    <script src="../js/cart.js" type="module"></script>
   </head>
 
   <body>
@@ -34,4 +33,5 @@
 
     <footer id="main-footer"></footer>
   </body>
+  <script src="../js/cart.js" type="module"></script>
 </html>
