+++ src/index.html
@@ -5,7 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Sleep Outside | Home</title>
     <link rel="stylesheet" href="/css/style.css" />
-    <script src="js/main.js" type="module"></script>
   </head>
 
   <body>
@@ -83,5 +82,7 @@
       </section>
     </main>
     <footer id="main-footer"></footer>
+    <script src="js/main.js" type="module"></script>
   </body>
+  
 </html>
