+++ src/js/productDetails.mjs
@@ -1,37 +1,21 @@
 import { findProductById } from "./productData.mjs";
-import { setLocalStorage, getLocalStorage } from "./utils.mjs";
+import {
+  setLocalStorage,
+  getLocalStorage,
+  renderSuperscriptNumbers,
+  loadHeaderFooter,
+} from "./utils.mjs";
 
 let product = {};
 
+loadHeaderFooter();
 export default async function productDetails(productId) {
-  try {
-    // get the details for the current product. findProductById will return a promise!
-    product = await findProductById(productId);
-    
-    // Check if the product is found
-    if (product) {
-      // once we have the product details we can render out the HTML
-      renderProductDetails();
-      // once the HTML is rendered we can add a listener to Add to Cart button
-      document.getElementById("addToCart").addEventListener("click", addToCart);
-
-      // Hide the error message when a product is found
-      displayProductNotFoundError(false);
-      // Show the Add to Cart button when a product is found
-      document.getElementById("addToCart").style.display = "block";
-    } else {
-      // Handle the case where the product is not found
-      displayProductNotFoundError(true);
-      // Hide the Add to Cart button when no product is found
-      document.getElementById("addToCart").style.display = "none";
-    }
-  } catch (error) {
-    // Handle any other errors that might occur during the API request
-    console.error("Error fetching product details:", error);
-    displayError("An error occurred while fetching product details.");
-    // Hide the Add to Cart button in case of an error
-    document.getElementById("addToCart").style.display = "none";
-  }
+  // get the details for the current product. findProductById will return a promise! use await or .then() to process it
+  product = await findProductById(productId);
+  // once we have the product details we can render out the HTML
+  renderProductDetails();
+  // once the HTML is rendered we can add a listener to Add to Cart button
+  document.getElementById("addToCart").addEventListener("click", addToCart);
 }
 
 function addToCart() {
@@ -40,9 +24,10 @@ function addToCart() {
   const cart = Array.isArray(storedCart) ? storedCart : [];
 
   cart.push(product);
-
   setLocalStorage("so-cart", cart);
+  renderSuperscriptNumbers();
 }
+
 function renderProductDetails() {
   document.querySelector("#productName").innerText = product.Brand.Name;
   document.querySelector("#productNameWithoutBrand").innerText =
@@ -56,19 +41,17 @@ function renderProductDetails() {
     product.DescriptionHtmlSimple;
   document.querySelector("#addToCart").dataset.id = product.Id;
 }
-function displayProductNotFoundError(showError) {
-  // Display or hide the error message based on the 'showError' parameter
-  const errorMessageElement = document.getElementById("errorMessage");
-  errorMessageElement.innerText = "Product not found.";
-  errorMessageElement.style.display = showError ? "block" : "none";
-}
+// function displayProductNotFoundError(showError) {
+//   // Display or hide the error message based on the 'showError' parameter
+//   const errorMessageElement = document.getElementById("errorMessage");
+//   errorMessageElement.innerText = "Product not found.";
+//   errorMessageElement.style.display = showError ? "block" : "none";
+// }
 
-
-function displayError(errorMessage) {
-  // Display a general error message
-  // show the error message
-  const errorMessageElement = document.getElementById("errorMessage");
-  errorMessageElement.innerText = errorMessage;
-  errorMessageElement.style.display = "block";
-}
-
+// function displayError(errorMessage) {
+//   // Display a general error message
+//   // show the error message
+//   const errorMessageElement = document.getElementById("errorMessage");
+//   errorMessageElement.innerText = errorMessage;
+//   errorMessageElement.style.display = "block";
+// }
