+++ src/public/partials/header.html
@@ -4,6 +4,7 @@
     <a href="/index.html"> Sleep<span class="highlight">Outside</span></a>
   </div>
   <div class="cart">
+    <span id="cart-count"></span>
     <a href="/cart/index.html">
       <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 100">
         <path
@@ -28,6 +29,7 @@
         <!-- <text x="0" y="115" fill="#000000" font-size="5px" font-weight="bold" font-family="'Helvetica Neue', Helvetica, Arial-Unicode, Arial, Sans-serif">Created by Natalia Woodroffe</text>
           <text x="0" y="120" fill="#000000" font-size="5px" font-weight="bold" font-family="'Helvetica Neue', Helvetica, Arial-Unicode, Arial, Sans-serif">from the Noun Project</text> -->
       </svg>
+      
     </a>
   </div>
